Magiclink/requirements.txt
fastapi>=0.103
uvicorn[standard]
pydantic>=2.0
python-jose[cryptography]
redis>=5.0
httpx>=0.27 